Refactor DynamicTable component and import OrderForUserTable in AllOrders
